156.Selectors
Created user.action, user.seelctor for selecting required state, declared user.types for optimization,  added Provider for state in index.js
