+++ crwn-clothing/src/index.js
@@ -4,22 +4,19 @@ import { BrowserRouter } from "react-router-dom";
 import "./index.scss";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
-import { UserProvider } from "./contexts/user.context";
 import { CartProvider } from "./contexts/cart.context";
 import { Provider } from "react-redux";
-import {store} from './store/store.js'
+import { store } from "./store/store.js";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
-  <React.StrictMode >
+  <React.StrictMode>
     <Provider store={store}>
-    <BrowserRouter>
-      <UserProvider>
-          <CartProvider>
-            <App />
-          </CartProvider>
-      </UserProvider>
-    </BrowserRouter>
+      <BrowserRouter>
+        <CartProvider>
+          <App />
+        </CartProvider>
+      </BrowserRouter>
     </Provider>
   </React.StrictMode>
 );
