+++ crwn-clothing/src/routes/navigation/navigation.component.jsx
@@ -1,15 +1,15 @@
-import { Link, Outlet } from "react-router-dom";
+import { Outlet } from "react-router-dom";
 import { Fragment, useContext } from "react";
-import { UserContext } from "../../contexts/user.context";
 import { signOutUser } from "../../utils/firebase/firebase.utils";
 import {NavLink, NavLinksContainer, NavigationContainer, LogoContainer} from "./navigation.styles";
-
+import { useSelector } from "react-redux";
 import CartDropdown from "../../components/cart-dropdown/cart-dropdown.component";
 import { ReactComponent as CrwnLogo } from "../../assets/crown.svg";
 import CartIcon from "../../components/cart-icon/cart-icon.component";
 import { CartContext } from "../../contexts/cart.context";
+import {selectCurrentUser} from '../../store/user/user.selector';
 const Navigation = () => {
-  const { currentUser } = useContext(UserContext);
+  const currentUser = useSelector(selectCurrentUser);
   const {isCartOpen} = useContext(CartContext);
   return (
     <Fragment>
