+++ crwn-clothing/src/store/root-reducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from "redux";
 import {userReducer} from './user/user.reducer';
 export const rootReducer = combineReducers({
-  // Add your reducers here
+  user: userReducer,
 });
