+++ crwn-clothing/src/store/user/user.selector.js
@@ -1,0 +1,1 @@
+export const selectCurrentUser = (state) => state.user.currentUser;
