+++ crwn-clothing/src/store/user/user.types.js
@@ -1,0 +1,3 @@
+export const USER_ACTION_TYPES = {
+    'SET_CURRENT_USER' : 'SET_CURRENT_USER'
+}
